@@ -1,8 +1,9 @@
-package com.vlath.keyboard;
-
 /**
- * Created by Vlad on 6/14/2017.
- */
+ * Copyright (C) 2017-2019 Vlad Todosin
+*/
+
+
+package com.vlath.keyboard;
 
 import android.graphics.Bitmap;
 import android.graphics.ColorMatrixColorFilter;
@@ -10,7 +11,6 @@ import android.graphics.drawable.ColorDrawable;
 import android.graphics.drawable.Drawable;
 import android.graphics.drawable.NinePatchDrawable;
 import android.inputmethodservice.KeyboardView;
-
 import android.content.Context;
 import android.graphics.Canvas;
 import android.graphics.Color;
@@ -33,20 +33,20 @@ public class CustomKeyboard extends KeyboardView {
 
     Drawable mTransparent = new ColorDrawable(Color.TRANSPARENT);
     NinePatchDrawable mSpaceBackground = (NinePatchDrawable) getContext().getResources().getDrawable(R.drawable.space);
-    NinePatchDrawable mPressedBackground = (NinePatchDrawable) getContext().getResources().getDrawable(R.drawable.press);
+    NinePatchDrawable mPressedBackground = (NinePatchDrawable) getContext().getResources()
+            .getDrawable(R.drawable.press);
     Paint mPaint = new Paint();
 
     public CustomKeyboard(Context context, AttributeSet attrs) {
         super(context, attrs);
     }
 
-
     public CustomKeyboard(Context context, AttributeSet attrs, int defStyle) {
         super(context, attrs, defStyle);
     }
 
-    public LatinKeyboard getLatinKeyboard(){
-        return (LatinKeyboard)getKeyboard();
+    public LatinKeyboard getLatinKeyboard() {
+        return (LatinKeyboard) getKeyboard();
     }
 
     @Override
@@ -67,9 +67,9 @@ public class CustomKeyboard extends KeyboardView {
 
         List<Key> keys = getKeyboard().getKeys();
 
-        for(Key key: keys) {
+        for (Key key : keys) {
 
-            if(key.label != null) {
+            if (key.label != null) {
                 if (key.codes[0] == 32) {
                     mSpaceBackground.setBounds(key.x, key.y, key.x + key.width, key.y + key.height);
                     mSpaceBackground.draw(canvas);
@@ -81,20 +81,19 @@ public class CustomKeyboard extends KeyboardView {
                             mPressedBackground.draw(canvas);
                         }
                     }
-                    if (Variables.isAlt()){
+                    if (Variables.isAlt()) {
                         if (key.codes[0] == -114) {
                             mPressedBackground.setBounds(key.x, key.y, key.x + key.width, key.y + key.height);
                             mPressedBackground.draw(canvas);
                         }
                     }
 
-                }
-                else{
-                    if(key.codes[0] == -113) {
+                } else {
+                    if (key.codes[0] == -113) {
                         mTransparent.setBounds(key.x, key.y, key.x + key.width, key.y + key.height);
                         mTransparent.draw(canvas);
                     }
-                    if(key.codes[0] == -114) {
+                    if (key.codes[0] == -114) {
                         mTransparent.setBounds(key.x, key.y, key.x + key.width, key.y + key.height);
                         mTransparent.draw(canvas);
                     }
